More work on unit details file transfer mechanism. /David & Jonas

+++ HsaTools/trunk/HRIV/modules/intsvc/src/main/java/se/vgregion/kivtools/hriv/intsvc/ws/eniro/FtpClientImpl.java
@@ -1,13 +1,19 @@
 package se.vgregion.kivtools.hriv.intsvc.ws.eniro;
 
 import java.io.ByteArrayInputStream;
+import java.io.IOException;
 import java.io.InputStream;
+import java.util.Date;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.commons.net.ftp.FTPClient;
 
+import se.vgregion.kivtools.search.svc.impl.kiv.ldap.Constants;
 
+/**
+ * Ftp client used for uploading files to ftp server.
+ */
 public class FtpClientImpl implements FtpClient {
 
   private Log logger = LogFactory.getLog(this.getClass());
@@ -42,25 +48,35 @@ public class FtpClientImpl implements FtpClient {
     this.ftpDestinationFileName = ftpDestinationFileName;
   }
 
+  /**
+   * {@inheritDoc}
+   */
   public boolean sendFile(String fileContent) {
+    if (fileContent == null) {
+      throw new IllegalArgumentException("Input string \"fileContent\" is null.");
+    }
+    boolean success;
     try {
       InputStream inputStream = new ByteArrayInputStream(fileContent.getBytes("UTF-8"));
       ftpclient.connect(hostname, port);
       ftpclient.enterLocalPassiveMode();
       boolean loginSuccess = ftpclient.login(username, password);
-      logger.debug("Unit details pusher: FTP login status: " + loginSuccess +". Server reply: " + ftpclient.getReplyString());
-      // Try to remove old file
+      logger.debug("Unit details pusher: FTP login status: " + loginSuccess + ". Server reply: " + ftpclient.getReplyString());
+      // Try to remove old file. Deletion should not be necessary but try to just in case something went wrong.
       boolean deleteSuccess = ftpclient.deleteFile(ftpDestinationFileName);
       if (deleteSuccess) {
-    	  logger.debug("Unit details pusher: Deleted " + ftpDestinationFileName + " on server.");
+        logger.debug("Unit details pusher: Deleted " + ftpDestinationFileName + " on server.");
       }
-      boolean storedSuccess = ftpclient.storeFile(ftpDestinationFileName, inputStream);
+      success = ftpclient.storeFile(ftpDestinationFileName + "-uploading.xml", inputStream);
+      if (success) {
+        success = ftpclient.rename(ftpDestinationFileName + "-uploading.xml", ftpDestinationFileName + "-" + Constants.formatDateToScientificTime(new Date()) + ".xml");
+      }
       inputStream.close();
       // Logout from the FTP Server and disconnect
       ftpclient.logout();
       ftpclient.disconnect();
-      return storedSuccess;
-    } catch (Exception e) {
+      return success;
+    } catch (IOException e) {
       logger.error("Error in SftpClient", e);
     }
     return false;
